Move downloads package to api/shared to facilitate its use by unichem
