+++ src/glados/api/shared/dynamic_downloads/tests/test_downloads_job_service.py
@@ -1,9 +1,9 @@
 from django.test import TestCase, override_settings
-from glados.api.chembl.dynamic_downloads.models import DownloadJob
+from glados.api.shared.dynamic_downloads.models import DownloadJob
 import json
 import os
 from django.conf import settings
-from glados.api.chembl.dynamic_downloads import download_job_service
+from glados.api.shared.dynamic_downloads import download_job_service
 import glados.es.ws2es.es_util as es_util
 from glados.settings import RunEnvs
 from datetime import timezone
