+++ src/glados/api/shared/dynamic_downloads/urls.py
@@ -1,4 +1,5 @@
 from django.conf.urls import url
+
 from . import downloads_controller
 
 urlpatterns = [
