+++ src/glados/api/shared/urls.py
@@ -1,0 +1,4 @@
+from django.conf.urls import include, url
+urlpatterns = [
+    url(r'downloads/', include('glados.api.shared.dynamic_downloads.urls')),
+]
